+++ app/services/db-export/db-export.service.js
@@ -5,10 +5,10 @@
  * @module DbExportService
  */
 
-const SchemaExportService = require('../db-export/schema-export.service')
+const SchemaExportService = require('../db-export/schema-export.service.js')
 
 /**
- * Calls TableExportsService giving it a tableName and schemaName
+ * Calls SchemaExportService giving it a schemaName
  */
 async function go () {
   const schemaNames = ['water', 'returns', 'crm', 'crm_v2', 'idm', 'permit']
