+++ app/services/db-export/schema-export.service.js
@@ -8,17 +8,23 @@
 const path = require('path')
 const os = require('os')
 
-const DeleteFolderService = require('./delete-folder.service')
-const FetchTableNames = require('../db-export/fetch-table-names.service')
-const TableExportService = require('../db-export/table-export.service')
-const SendToS3BucketService = require('../db-export/send-to-s3-bucket.service')
+const DeleteFolderService = require('./delete-folder.service.js')
+const FetchTableNames = require('../db-export/fetch-table-names.service.js')
+const SendToS3BucketService = require('../db-export/send-to-s3-bucket.service.js')
+const ExportCompressedTableService = require('./export-compressed-table.service.js')
 
+/**
+ * Exports the specific schema by fetching table names, exporting each table,
+ * and uploading the schema folder to an S3 bucket
+ *
+ * @param {String} schemaName The name of the database to export
+ */
 async function go (schemaName) {
   const tableNames = await FetchTableNames.go(schemaName)
 
   const schemaFolderPath = _folderToUpload(schemaName)
   for (const tableName of tableNames) {
-    await TableExportService.go(tableName, schemaFolderPath, schemaName)
+    await ExportCompressedTableService.go(tableName, schemaFolderPath, schemaName)
   }
 
   await SendToS3BucketService.go(schemaFolderPath)
@@ -26,6 +32,13 @@ async function go (schemaName) {
   await DeleteFolderService.go(schemaFolderPath)
 }
 
+/**
+ * Generates the folder path where the schema will be temporarily stored for upload
+ *
+ * @param {String} schemaName The name of the database schema
+ *
+ * @returns {String} The folder path where the schema will be temporarily stored
+ */
 function _folderToUpload (schemaName) {
   const temporaryFilePath = os.tmpdir()
